+++ app/frontends/angularjs/src/main/scala/com/shocktrade/client/users/HomeController.scala
@@ -83,7 +83,7 @@ object HomeController {
     var getLevelDescription: js.Function0[js.UndefOr[String]] = js.native
     var getNextLevelXP: js.Function0[js.UndefOr[Int]] = js.native
     var getStars: js.Function0[js.UndefOr[js.Array[Int]]] = js.native
-    var getTotalXP: js.Function0[js.UndefOr[Int]] = js.native
+    var getTotalXP: js.Function0[js.UndefOr[Double]] = js.native
   }
 
 }
